Should fix #12
This little update has been made to fix a rare bug where a Lua error appear when a player spawn for the first time in the server.

I'm not sure this will fix the issue, but after few searches, I've found that sometimes, the players isn't totally loaded so the "ply" object isn't yet initialized. That's why I made a verification with IsValid() and IsPlayer().

Thanks.